# app/controllers.go
// Code generated by goagen v1.3.0, DO NOT EDIT.
//
// API "fabric8-starter": Application Controllers
//
// Command:
// $ goagen
// --design=github.com/fabric8-services/fabric8-starter/design
// --out=$(GOPATH)/src/github.com/fabric8-services/fabric8-starter
// --version=v1.3.0

package app

import (
	"context"
	"github.com/goadesign/goa"
	"github.com/goadesign/goa/cors"
	"net/http"
	"regexp"
)

// initService sets up the service encoders, decoders and mux.
func initService(service *goa.Service) {
	// Setup encoders and decoders
	service.Encoder.Register(goa.NewJSONEncoder, "application/json")
	service.Decoder.Register(goa.NewJSONDecoder, "application/json")

	// Setup default encoder and decoder
	service.Encoder.Register(goa.NewJSONEncoder, "*/*")
	service.Decoder.Register(goa.NewJSONDecoder, "*/*")
}

// LabelController is the controller interface for the Label actions.
type LabelController interface {
	goa.Muxer
	Create(*CreateLabelContext) error
	List(*ListLabelContext) error
	Show(*ShowLabelContext) error
	Update(*UpdateLabelContext) error
}

// MountLabelController "mounts" a Label resource controller on the given service.
func MountLabelController(service *goa.Service, ctrl LabelController) {
	initService(service)
	var h goa.Handler
	service.Mux.Handle("OPTIONS", "/api/labels", ctrl.MuxHandler("preflight", handleLabelOrigin(cors.HandlePreflight()), nil))
	service.Mux.Handle("OPTIONS", "/api/labels/:labelID", ctrl.MuxHandler("preflight", handleLabelOrigin(cors.HandlePreflight()), nil))

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewCreateLabelContext(ctx, req, service)
		if err != nil {
			return err
		}
		// Build the payload
		if rawPayload := goa.ContextRequest(ctx).Payload; rawPayload != nil {
			rctx.Payload = rawPayload.(*CreateLabelPayload)
		} else {
			return goa.MissingPayloadError()
		}
		return ctrl.Create(rctx)
	}
	h = handleSecurity("jwt", h)
	h = handleLabelOrigin(h)
	service.Mux.Handle("POST", "/api/labels", ctrl.MuxHandler("create", h, unmarshalCreateLabelPayload))
	service.LogInfo("mount", "ctrl", "Label", "action", "Create", "route", "POST /api/labels", "security", "jwt")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewListLabelContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.List(rctx)
	}
	h = handleLabelOrigin(h)
	service.Mux.Handle("GET", "/api/labels", ctrl.MuxHandler("list", h, nil))
	service.LogInfo("mount", "ctrl", "Label", "action", "List", "route", "GET /api/labels")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewShowLabelContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Show(rctx)
	}
	h = handleLabelOrigin(h)
	service.Mux.Handle("GET", "/api/labels/:labelID", ctrl.MuxHandler("show", h, nil))
	service.LogInfo("mount", "ctrl", "Label", "action", "Show", "route", "GET /api/labels/:labelID")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewUpdateLabelContext(ctx, req, service)
		if err != nil {
			return err
		}
		// Build the payload
		if rawPayload := goa.ContextRequest(ctx).Payload; rawPayload != nil {
			rctx.Payload = rawPayload.(*UpdateLabelPayload)
		} else {
			return goa.MissingPayloadError()
		}
		return ctrl.Update(rctx)
	}
	h = handleSecurity("jwt", h)
	h = handleLabelOrigin(h)
	service.Mux.Handle("PATCH", "/api/labels/:labelID", ctrl.MuxHandler("update", h, unmarshalUpdateLabelPayload))
	service.LogInfo("mount", "ctrl", "Label", "action", "Update", "route", "PATCH /api/labels/:labelID", "security", "jwt")
}

// handleLabelOrigin applies the CORS response headers corresponding to the origin.
func handleLabelOrigin(h goa.Handler) goa.Handler {
	spec0 := regexp.MustCompile("[.*openshift.io|localhost]")

	return func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		origin := req.Header.Get("Origin")
		if origin == "" {
			// Not a CORS request
			return h(ctx, rw, req)
		}
		if cors.MatchOriginRegexp(origin, spec0) {
			ctx = goa.WithLogContext(ctx, "origin", origin)
			rw.Header().Set("Access-Control-Allow-Origin", origin)
			rw.Header().Set("Vary", "Origin")
			rw.Header().Set("Access-Control-Max-Age", "600")
			rw.Header().Set("Access-Control-Allow-Credentials", "true")
			if acrm := req.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				rw.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT, PATCH, DELETE")
				rw.Header().Set("Access-Control-Allow-Headers", "X-Request-Id, Content-Type, Authorization, If-None-Match, If-Modified-Since")
			}
			return h(ctx, rw, req)
		}

		return h(ctx, rw, req)
	}
}

// unmarshalCreateLabelPayload unmarshals the request body into the context request data Payload field.
func unmarshalCreateLabelPayload(ctx context.Context, service *goa.Service, req *http.Request) error {
	payload := &createLabelPayload{}
	if err := service.DecodeRequest(req, payload); err != nil {
		return err
	}
	if err := payload.Validate(); err != nil {
		// Initialize payload with private data structure so it can be logged
		goa.ContextRequest(ctx).Payload = payload
		return err
	}
	goa.ContextRequest(ctx).Payload = payload.Publicize()
	return nil
}

// unmarshalUpdateLabelPayload unmarshals the request body into the context request data Payload field.
func unmarshalUpdateLabelPayload(ctx context.Context, service *goa.Service, req *http.Request) error {
	payload := &updateLabelPayload{}
	if err := service.DecodeRequest(req, payload); err != nil {
		return err
	}
	if err := payload.Validate(); err != nil {
		// Initialize payload with private data structure so it can be logged
		goa.ContextRequest(ctx).Payload = payload
		return err
	}
	goa.ContextRequest(ctx).Payload = payload.Publicize()
	return nil
}

// StatusController is the controller interface for the Status actions.
type StatusController interface {
	goa.Muxer
	Show(*ShowStatusContext) error
}

// MountStatusController "mounts" a Status resource controller on the given service.
func MountStatusController(service *goa.Service, ctrl StatusController) {
	initService(service)
	var h goa.Handler
	service.Mux.Handle("OPTIONS", "/api/status", ctrl.MuxHandler("preflight", handleStatusOrigin(cors.HandlePreflight()), nil))

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewShowStatusContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Show(rctx)
	}
	h = handleStatusOrigin(h)
	service.Mux.Handle("GET", "/api/status", ctrl.MuxHandler("show", h, nil))
	service.LogInfo("mount", "ctrl", "Status", "action", "Show", "route", "GET /api/status")
}

// handleStatusOrigin applies the CORS response headers corresponding to the origin.
func handleStatusOrigin(h goa.Handler) goa.Handler {
	spec0 := regexp.MustCompile("[.*openshift.io|localhost]")

	return func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		origin := req.Header.Get("Origin")
		if origin == "" {
			// Not a CORS request
			return h(ctx, rw, req)
		}
		if cors.MatchOriginRegexp(origin, spec0) {
			ctx = goa.WithLogContext(ctx, "origin", origin)
			rw.Header().Set("Access-Control-Allow-Origin", origin)
			rw.Header().Set("Vary", "Origin")
			rw.Header().Set("Access-Control-Max-Age", "600")
			rw.Header().Set("Access-Control-Allow-Credentials", "true")
			if acrm := req.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				rw.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT, PATCH, DELETE")
				rw.Header().Set("Access-Control-Allow-Headers", "X-Request-Id, Content-Type, Authorization, If-None-Match, If-Modified-Since")
			}
			return h(ctx, rw, req)
		}

		return h(ctx, rw, req)
	}
}