# swagger/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
  CreateLabelPayload:
    description: Holds a single Label
    example:
      data:
        attributes:
          background-color: '#ffa7cb'
          border-color: '#ffa7cb'
          created-at: "2016-11-29T23:18:14Z"
          name: name for the object
          text-color: '#ffa7cb'
          updated-at: "2016-11-29T23:18:14Z"
          version: 23
        id: 40bbdd3d-8b5d-4fd6-ac90-7236b669af04
        links:
          meta:
            Veritatis non aut quis tempore velit ducimus.: 0.15907161337191197
          related: Sapiente nesciunt est soluta eos.
          self: Laborum laudantium reprehenderit maiores nulla ut.
        relationships:
          space:
            data:
              id: 6c5610be-30b2-4880-9fec-81e4f8e4fd76
              links:
                meta:
                  Veritatis non aut quis tempore velit ducimus.: 0.15907161337191197
                related: Sapiente nesciunt est soluta eos.
                self: Laborum laudantium reprehenderit maiores nulla ut.
              type: Dignissimos iste recusandae iure dolorem odit.
            links:
              meta:
                Veritatis non aut quis tempore velit ducimus.: 0.15907161337191197
              related: Sapiente nesciunt est soluta eos.
              self: Laborum laudantium reprehenderit maiores nulla ut.
            meta:
              Voluptas iste corrupti labore.: 0.2152925220749911
        type: labels
      included:
      - "1979-05-01T16:19:31Z"
    properties:
      data:
        $ref: '#/definitions/Label'
      included:
        description: An array of mixed types
        example:
        - "1979-05-01T16:19:31Z"
        items:
          example: "1979-05-01T16:19:31Z"
        type: array
    required:
    - data
    title: CreateLabelPayload
    type: object
...
  GenericLinks:
    example:
      meta:
        Veritatis non aut quis tempore velit ducimus.: 0.15907161337191197
      related: Sapiente nesciunt est soluta eos.
      self: Laborum laudantium reprehenderit maiores nulla ut.
    properties:
      meta:
        additionalProperties: true
        example:
          Veritatis non aut quis tempore velit ducimus.: 0.15907161337191197
        type: object
      related:
        example: Sapiente nesciunt est soluta eos.
        type: string
      self:
        example: Laborum laudantium reprehenderit maiores nulla ut.
        type: string
    title: GenericLinks
    type: object
  JSONAPIError:
    description: |-
      Error objects provide additional information about problems encountered while
      performing an operation. Error objects MUST be returned as an array keyed by errors in the
      top level of a JSON API document.

      See. also http://jsonapi.org/format/#error-objects.
    example:
      code: Rerum natus.
      detail: Quisquam magni itaque quia nulla minima.
      id: Cupiditate enim architecto magnam ut autem.
      links:
        Iure praesentium ipsum recusandae.:
          href: http://example.com/articles/1/comments
          meta:
            Rerum laborum suscipit nam voluptate.: "1990-10-14T23:24:35Z"
      meta:
        Ut temporibus.: "1976-10-10T14:09:07Z"
      source:
        Eos non rerum.: 0.3968728190078231
      status: Ea quo voluptatem ipsa.
      title: Sapiente enim inventore.
    properties:
      code:
        description: an application-specific error code, expressed as a string value.
        example: Rerum natus.
        type: string
      detail:
        description: |-
          a human-readable explanation specific to this occurrence of the problem.
          Like title, this field’s value can be localized.
        example: Quisquam magni itaque quia nulla minima.
        type: string
      id:
        description: a unique identifier for this particular occurrence of the problem.
        example: Cupiditate enim architecto magnam ut autem.
        type: string
      links:
        additionalProperties: true
        description: |-
          a links object containing the following members:
          * about: a link that leads to further details about this particular occurrence of the problem.
        example:
          Iure praesentium ipsum recusandae.:
            href: http://example.com/articles/1/comments
            meta:
              Rerum laborum suscipit nam voluptate.: "1990-10-14T23:24:35Z"
        type: object
      meta:
        additionalProperties: true
        description: a meta object containing non-standard meta-information about
          the error
        example:
          Ut temporibus.: "1976-10-10T14:09:07Z"
        type: object
      source:
        additionalProperties: true
        description: |-
          an object containing references to the source of the error,
          optionally including any of the following members

          * pointer: a JSON Pointer [RFC6901] to the associated entity in the request document [e.g. "/data" for a primary data object,
                     or "/data/attributes/title" for a specific attribute].
          * parameter: a string indicating which URI query parameter caused the error.
        example:
          Eos non rerum.: 0.3968728190078231
        type: object
      status:
        description: the HTTP status code applicable to this problem, expressed as
          a string value.
        example: Ea quo voluptatem ipsa.
        type: string
      title:
        description: |-
          a short, human-readable summary of the problem that SHOULD NOT
          change from occurrence to occurrence of the problem, except for purposes of localization.
        example: Sapiente enim inventore.
        type: string
    required:
    - detail
    title: JSONAPIError
    type: object
  JSONAPIErrors:
    description: JSONAPIErrors media type (default view)
    example:
      errors:
      - code: Rerum natus.
        detail: Quisquam magni itaque quia nulla minima.
        id: Cupiditate enim architecto magnam ut autem.
        links:
          Iure praesentium ipsum recusandae.:
            href: http://example.com/articles/1/comments
            meta:
              Rerum laborum suscipit nam voluptate.: "1990-10-14T23:24:35Z"
        meta:
          Ut temporibus.: "1976-10-10T14:09:07Z"
        source:
          Eos non rerum.: 0.3968728190078231
        status: Ea quo voluptatem ipsa.
        title: Sapiente enim inventore.
    properties:
      errors:
        example:
        - code: Rerum natus.
          detail: Quisquam magni itaque quia nulla minima.
          id: Cupiditate enim architecto magnam ut autem.
          links:
            Iure praesentium ipsum recusandae.:
              href: http://example.com/articles/1/comments
              meta:
                Rerum laborum suscipit nam voluptate.: "1990-10-14T23:24:35Z"
          meta:
            Ut temporibus.: "1976-10-10T14:09:07Z"
          source:
            Eos non rerum.: 0.3968728190078231
          status: Ea quo voluptatem ipsa.
          title: Sapiente enim inventore.
        items:
          $ref: '#/definitions/JSONAPIError'
        type: array
    required:
    - errors
    title: 'Mediatype identifier: application/vnd.jsonapierrors+json; view=default'
    type: object
  Label:
    description: JSONAPI store for the data of a Label. See also http://jsonapi.org/format/#document-resource-object
    example:
      attributes:
        background-color: '#ffa7cb'
        border-color: '#ffa7cb'
        created-at: "2016-11-29T23:18:14Z"
        name: name for the object
        text-color: '#ffa7cb'
        updated-at: "2016-11-29T23:18:14Z"
        version: 23
      id: 40bbdd3d-8b5d-4fd6-ac90-7236b669af04
      links:
        meta:
          Veritatis non aut quis tempore velit ducimus.: 0.15907161337191197
        related: Sapiente nesciunt est soluta eos.
        self: Laborum laudantium reprehenderit maiores nulla ut.
      relationships:
        space:
          data:
            id: 6c5610be-30b2-4880-9fec-81e4f8e4fd76
            links:
              meta:
                Veritatis non aut quis tempore velit ducimus.: 0.15907161337191197
              related: Sapiente nesciunt est soluta eos.
              self: Laborum laudantium reprehenderit maiores nulla ut.
            type: Dignissimos iste recusandae iure dolorem odit.
          links:
            meta:
              Veritatis non aut quis tempore velit ducimus.: 0.15907161337191197
            related: Sapiente nesciunt est soluta eos.
            self: Laborum laudantium reprehenderit maiores nulla ut.
          meta:
            Voluptas iste corrupti labore.: 0.2152925220749911
      type: labels
    properties:
      attributes:
        $ref: '#/definitions/LabelAttributes'
      id:
        description: ID of label
        example: 40bbdd3d-8b5d-4fd6-ac90-7236b669af04
        format: uuid
        type: string
      links:
        $ref: '#/definitions/GenericLinks'
      relationships:
        $ref: '#/definitions/LabelRelations'
      type:
        enum:
        - labels
        example: labels
        type: string
    required:
    - type
    - attributes
    title: Label
    type: object
  LabelAttributes:
    description: JSONAPI store for all the "attributes" of a Label. See also http://jsonapi.org/format/#document-resource-object-attributes
    example:
      background-color: '#ffa7cb'
      border-color: '#ffa7cb'
      created-at: "2016-11-29T23:18:14Z"
      name: name for the object
      text-color: '#ffa7cb'
      updated-at: "2016-11-29T23:18:14Z"
      version: 23
    properties:
      background-color:
        description: Background color in hex code format. See also http://www.color-hex.com
        example: '#ffa7cb'
        type: string
      border-color:
        description: Border color in hex code format. See also http://www.color-hex.com
        example: '#ffa7cb'
        type: string
      created-at:
        description: When the label was created
        example: "2016-11-29T23:18:14Z"
        format: date-time
        type: string
      name:
        description: The Label name
        example: name for the object
        maxLength: 63
        minLength: 1
        pattern: ^[^_|-].*
        type: string
      text-color:
        description: Text color in hex code format. See also http://www.color-hex.com
        example: '#ffa7cb'
        type: string
      updated-at:
        description: When the label was updated
        example: "2016-11-29T23:18:14Z"
        format: date-time
        type: string
      version:
        description: Version for optimistic concurrency control (optional during creating)
        example: 23
        format: int64
        type: integer
    title: LabelAttributes
    type: object
  LabelList:
    description: Holds the list of Labels (default view)
    example:
      data:
      - attributes:
          background-color: '#ffa7cb'
          border-color: '#ffa7cb'
          created-at: "2016-11-29T23:18:14Z"
          name: name for the object
          text-color: '#ffa7cb'
          updated-at: "2016-11-29T23:18:14Z"
          version: 23
        id: 40bbdd3d-8b5d-4fd6-ac90-7236b669af04
        links:
          meta:
            Veritatis non aut quis tempore velit ducimus.: 0.15907161337191197
          related: Sapiente nesciunt est soluta eos.
          self: Laborum laudantium reprehenderit maiores nulla ut.
        relationships:
          space:
            data:
              id: 6c5610be-30b2-4880-9fec-81e4f8e4fd76
              links:
                meta:
                  Veritatis non aut quis tempore velit ducimus.: 0.15907161337191197
                related: Sapiente nesciunt est soluta eos.
                self: Laborum laudantium reprehenderit maiores nulla ut.
              type: Dignissimos iste recusandae iure dolorem odit.
            links:
              meta:
                Veritatis non aut quis tempore velit ducimus.: 0.15907161337191197
              related: Sapiente nesciunt est soluta eos.
              self: Laborum laudantium reprehenderit maiores nulla ut.
            meta:
              Voluptas iste corrupti labore.: 0.2152925220749911
        type: labels
      included:
      - "2000-03-27T17:16:30Z"
      - "2000-03-27T17:16:30Z"
      links:
        filters: Velit voluptatem sunt dolorem.
        first: Laboriosam quibusdam porro.
        last: Rem earum odio quis sunt velit.
        next: Magnam qui facilis non odio.
        prev: Voluptate iure est ut doloremque.
    properties:
      data:
        example:
        - attributes:
            background-color: '#ffa7cb'
            border-color: '#ffa7cb'
            created-at: "2016-11-29T23:18:14Z"
            name: name for the object
            text-color: '#ffa7cb'
            updated-at: "2016-11-29T23:18:14Z"
            version: 23
          id: 40bbdd3d-8b5d-4fd6-ac90-7236b669af04
          links:
            meta:
              Veritatis non aut quis tempore velit ducimus.: 0.15907161337191197
            related: Sapiente nesciunt est soluta eos.
            self: Laborum laudantium reprehenderit maiores nulla ut.
          relationships:
            space:
              data:
                id: 6c5610be-30b2-4880-9fec-81e4f8e4fd76
                links:
                  meta:
                    Veritatis non aut quis tempore velit ducimus.: 0.15907161337191197
                  related: Sapiente nesciunt est soluta eos.
                  self: Laborum laudantium reprehenderit maiores nulla ut.
                type: Dignissimos iste recusandae iure dolorem odit.
              links:
                meta:
                  Veritatis non aut quis tempore velit ducimus.: 0.15907161337191197
                related: Sapiente nesciunt est soluta eos.
                self: Laborum laudantium reprehenderit maiores nulla ut.
              meta:
                Voluptas iste corrupti labore.: 0.2152925220749911
          type: labels
        items:
          $ref: '#/definitions/Label'
        type: array
      included:
        description: An array of mixed types
        example:
        - "2000-03-27T17:16:30Z"
        - "2000-03-27T17:16:30Z"
        items:
          example: "2000-03-27T17:16:30Z"
        type: array
      links:
        $ref: '#/definitions/pagingLinks'
    required:
    - data
    title: 'Mediatype identifier: application/vnd.labellist+json; view=default'
    type: object
...
  LabelSingle:
    description: Holds a single Label (default view)
    example:
      data:
        attributes:
          background-color: '#ffa7cb'
          border-color: '#ffa7cb'
          created-at: "2016-11-29T23:18:14Z"
          name: name for the object
          text-color: '#ffa7cb'
          updated-at: "2016-11-29T23:18:14Z"
          version: 23
        id: 40bbdd3d-8b5d-4fd6-ac90-7236b669af04
        links:
          meta:
            Veritatis non aut quis tempore velit ducimus.: 0.15907161337191197
          related: Sapiente nesciunt est soluta eos.
          self: Laborum laudantium reprehenderit maiores nulla ut.
        relationships:
          space:
            data:
              id: 6c5610be-30b2-4880-9fec-81e4f8e4fd76
              links:
                meta:
                  Veritatis non aut quis tempore velit ducimus.: 0.15907161337191197
                related: Sapiente nesciunt est soluta eos.
                self: Laborum laudantium reprehenderit maiores nulla ut.
              type: Dignissimos iste recusandae iure dolorem odit.
            links:
              meta:
                Veritatis non aut quis tempore velit ducimus.: 0.15907161337191197
              related: Sapiente nesciunt est soluta eos.
              self: Laborum laudantium reprehenderit maiores nulla ut.
            meta:
              Voluptas iste corrupti labore.: 0.2152925220749911
        type: labels
      included:
      - "1979-05-01T16:19:31Z"
    properties:
      data:
        $ref: '#/definitions/Label'
      included:
        description: An array of mixed types
        example:
        - "1979-05-01T16:19:31Z"
        items:
          example: "1979-05-01T16:19:31Z"
        type: array
    required:
    - data
    title: 'Mediatype identifier: application/vnd.label+json; view=default'
    type: object
...
  Status:
    description: The status of the current running instance (default view)
    example:
      buildTime: Enim numquam nobis ut in accusantium delectus.
      commit: Laboriosam fugit numquam rerum.
      error: Unde suscipit.
      startTime: Praesentium vitae voluptatem.
    properties:
      buildTime:
        description: The time when built
        example: Enim numquam nobis ut in accusantium delectus.
        type: string
      commit:
        description: Commit SHA this build is based on
        example: Laboriosam fugit numquam rerum.
        type: string
      error:
        description: The error if any
        example: Unde suscipit.
        type: string
      startTime:
        description: The time when started
        example: Praesentium vitae voluptatem.
        type: string
    required:
    - commit
    - buildTime
    - startTime
    title: 'Mediatype identifier: application/vnd.status+json; view=default'
    type: object
  UpdateLabelPayload:
    description: Holds a single Label
    example:
      data:
        attributes:
          background-color: '#ffa7cb'
          border-color: '#ffa7cb'
          created-at: "2016-11-29T23:18:14Z"
          name: name for the object
          text-color: '#ffa7cb'
          updated-at: "2016-11-29T23:18:14Z"
          version: 23
        id: 40bbdd3d-8b5d-4fd6-ac90-7236b669af04
        links:
          meta:
            Veritatis non aut quis tempore velit ducimus.: 0.15907161337191197
          related: Sapiente nesciunt est soluta eos.
          self: Laborum laudantium reprehenderit maiores nulla ut.
        relationships:
          space:
            data:
              id: 6c5610be-30b2-4880-9fec-81e4f8e4fd76
              links:
                meta:
                  Veritatis non aut quis tempore velit ducimus.: 0.15907161337191197
                related: Sapiente nesciunt est soluta eos.
                self: Laborum laudantium reprehenderit maiores nulla ut.
              type: Dignissimos iste recusandae iure dolorem odit.
            links:
              meta:
                Veritatis non aut quis tempore velit ducimus.: 0.15907161337191197
              related: Sapiente nesciunt est soluta eos.
              self: Laborum laudantium reprehenderit maiores nulla ut.
            meta:
              Voluptas iste corrupti labore.: 0.2152925220749911
        type: labels
      included:
      - "1979-05-01T16:19:31Z"
    properties:
      data:
        $ref: '#/definitions/Label'
      included:
        description: An array of mixed types
        example:
        - "1979-05-01T16:19:31Z"
        items:
          example: "1979-05-01T16:19:31Z"
        type: array
    required:
    - data
    title: UpdateLabelPayload
    type: object
  pagingLinks:
    example:
      filters: Velit voluptatem sunt dolorem.
      first: Laboriosam quibusdam porro.
      last: Rem earum odio quis sunt velit.
      next: Magnam qui facilis non odio.
      prev: Voluptate iure est ut doloremque.
    properties:
      filters:
        example: Velit voluptatem sunt dolorem.
        type: string
      first:
        example: Laboriosam quibusdam porro.
        type: string
      last:
        example: Rem earum odio quis sunt velit.
        type: string
      next:
        example: Magnam qui facilis non odio.
        type: string
      prev:
        example: Voluptate iure est ut doloremque.
        type: string
    title: pagingLinks
    type: object
host: openshift.io
info:
  license:
    name: Apache License Version 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  version: ""
paths:
  /labels:
    get:
      description: List Labels.
      operationId: label#list
      parameters:
      - in: header
        name: If-Modified-Since
        required: false
        type: string
      - in: header
        name: If-None-Match
        required: false
        type: string
      produces:
      - application/vnd.jsonapierrors+json
      - application/vnd.jsonapierrors+json
      - application/vnd.labellist+json
      responses:
        "200":
          description: Resource created
          headers:
            Cache-Control:
              type: string
            ETag:
              type: string
            Last-Modified:
              type: string
          schema:
            $ref: '#/definitions/LabelList'
        "304":
          description: Not Modified
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/JSONAPIErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/JSONAPIErrors'
      schemes:
      - http
      summary: list label
      tags:
      - label
    post:
      description: create label with id, name and color.
      operationId: label#create
      parameters:
      - description: Holds a single Label
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/CreateLabelPayload'
      produces:
      - application/vnd.jsonapierrors+json
      - application/vnd.jsonapierrors+json
      - application/vnd.label+json
      - application/vnd.jsonapierrors+json
      - application/vnd.jsonapierrors+json
      - application/vnd.jsonapierrors+json
      - application/vnd.jsonapierrors+json
      responses:
        "201":
          description: Resource created
          headers:
            Location:
              description: href to created resource
              pattern: /labels/.*
              type: string
          schema:
            $ref: '#/definitions/LabelSingle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/JSONAPIErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/JSONAPIErrors'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/JSONAPIErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/JSONAPIErrors'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/JSONAPIErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/JSONAPIErrors'
      schemes:
      - http
      security:
      - jwt: []
      summary: create label
      tags:
      - label
  /labels/{labelID}:
    get:
      description: Retrieve label for the given id.
      operationId: label#show
      parameters:
      - description: ID of the label
        in: path
        name: labelID
        required: true
        type: string
      - in: header
        name: If-Modified-Since
        required: false
        type: string
      - in: header
        name: If-None-Match
        required: false
        type: string
      produces:
      - application/vnd.jsonapierrors+json
      - application/vnd.jsonapierrors+json
      - application/vnd.jsonapierrors+json
      - application/vnd.label+json
      responses:
        "200":
          description: Resource created
          headers:
            Cache-Control:
              type: string
            ETag:
              type: string
            Last-Modified:
              type: string
          schema:
            $ref: '#/definitions/LabelSingle'
        "304":
          description: Not Modified
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/JSONAPIErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/JSONAPIErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/JSONAPIErrors'
      schemes:
      - http
      summary: show label
      tags:
      - label
    patch:
      description: update the label for the given id.
      operationId: label#update
      parameters:
      - description: ID of the label to update
        in: path
        name: labelID
        required: true
        type: string
      - description: Holds a single Label
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/UpdateLabelPayload'
      produces:
      - application/vnd.jsonapierrors+json
      - application/vnd.jsonapierrors+json
      - application/vnd.jsonapierrors+json
      - application/vnd.jsonapierrors+json
      - application/vnd.jsonapierrors+json
      - application/vnd.label+json
      - application/vnd.jsonapierrors+json
      responses:
        "200":
          description: Resource created
          headers:
            Cache-Control:
              type: string
            ETag:
              type: string
            Last-Modified:
              type: string
          schema:
            $ref: '#/definitions/LabelSingle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/JSONAPIErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/JSONAPIErrors'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/JSONAPIErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/JSONAPIErrors'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/JSONAPIErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/JSONAPIErrors'
      schemes:
      - http
      security:
      - jwt: []
      summary: update label
      tags:
      - label
  /status:
    get:
      description: Show the status of the current running instance
      operationId: status#show
      produces:
      - application/vnd.status+json
      - application/vnd.status+json
      responses:
        "200":
          description: Resource created
          headers:
            Cache-Control:
              type: string
            ETag:
              type: string
            Last-Modified:
              type: string
          schema:
            $ref: '#/definitions/Status'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/Status'
      schemes:
      - http
      summary: show status
      tags:
      - status
produces:
- application/json
...
schemes:
- http
securityDefinitions:
  jwt:
    description: |-
      JWT Token Auth

      **Token URL**: http://openshift.io/api/login/authorize
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"